+++ CDatos/FASV1_GetChangesOnProduct_Desc_Result.cs
@@ -15,6 +15,6 @@ namespace CDatos
     {
         public int IdLog { get; set; }
         public int IdProduct { get; set; }
-        public byte ActionMade { get; set; }
+        public int ActionMade { get; set; }
     }
 }
